+++ projects/athena-numeric/src/real.rs
@@ -1,6 +1,6 @@
-//! Real numbers: IEEE binary64 or binary [`BigFloat`].
+//! Real numbers: IEEE binary64 or binary [`Decimal`].
 
-use crate::big_float::BigFloat;
+use crate::decimal::Decimal;
 
 /// Real value representation.
 #[derive(Debug, Clone, PartialEq)]
@@ -8,7 +8,7 @@ pub enum Real {
     /// IEEE binary64 (includes non-finite values).
     Machine(f64),
     /// Finite binary float with explicit working precision.
-    BigFloat(BigFloat),
+    BigFloat(Decimal),
 }
 
 impl Real {
@@ -18,11 +18,11 @@ impl Real {
     }
 
     /// Arbitrary-precision finite real.
-    pub fn big_float(b: BigFloat) -> Self {
+    pub fn big_float(b: Decimal) -> Self {
         Self::BigFloat(b)
     }
 
-    /// Machine `f64` view (exact for [`Self::Machine`], exact when [`BigFloat::to_f64_exact`] succeeds).
+    /// Machine `f64` view (exact for [`Self::Machine`], exact when [`Decimal::to_f64_exact`] succeeds).
     pub fn as_f64(&self) -> Option<f64> {
         match self {
             Self::Machine(x) => Some(*x),
@@ -38,8 +38,8 @@ impl Real {
         }
     }
 
-    /// View as [`BigFloat`] when already arbitrary.
-    pub fn as_big_float(&self) -> Option<&BigFloat> {
+    /// View as [`Decimal`] when already arbitrary.
+    pub fn as_big_float(&self) -> Option<&Decimal> {
         match self {
             Self::BigFloat(b) => Some(b),
             _ => None,
